+++ weblate/formats/defaults.py
@@ -15,6 +15,7 @@ DEFAULT_FORMATS: tuple[str, ...] = (
     "weblate.formats.ttkit.PoXliffFormat",
     "weblate.formats.ttkit.AppleXliffFormat",
     "weblate.formats.ttkit.StringsFormat",
+    "weblate.formats.ttkit.XCStringsFormat",
     "weblate.formats.ttkit.PropertiesFormat",
     "weblate.formats.ttkit.JoomlaFormat",
     "weblate.formats.ttkit.GWTFormat",